Attempting to add in navigation on home screen, nothing stupid broken yet
