--- yekyok/home_screens/MapScreen.js
@@ -1,1 +1,0 @@
-export default class MapScreen extends React.Component {}
