--- yekyok/home_screens/NewScreen.js
@@ -1,1 +1,0 @@
-export default class NewScreen extends React.Component {}
