+++ yekyok/navigation/AppNavigator.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 
 import MainTabNavigator from './MainTabNavigator';
