+++ yekyok/navigation/MainTabNavigator.js
@@ -3,13 +3,17 @@ import { Platform } from 'react-native';
 import { createStackNavigator, createBottomTabNavigator } from 'react-navigation';
 
 import TabBarIcon from '../components/TabBarIcon';
-import HomeScreen from '../screens/HomeScreen';
 import LinksScreen from '../screens/LinksScreen';
 import SettingsScreen from '../screens/SettingsScreen';
 import ProfileScreen from '../screens/ProfileScreen';
+import MapScreen from '../screens/home_screens/MapScreen';
+import NewScreen from '../screens/home_screens/NewScreen';
+import HotScreen from '../screens/home_screens/HotScreen';
 
 const HomeStack = createStackNavigator({
-  Home: HomeScreen,
+  Map: MapScreen,
+  New: NewScreen,
+  Hot: HotScreen
 });
 
 HomeStack.navigationOptions = {
@@ -39,7 +43,6 @@ LinksStack.navigationOptions = {
     />
   ),
 };
-
 
 const ProfileStack = createStackNavigator({
   Profile: ProfileScreen,
