+++ yekyok/screens/home_screens/HotScreen.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+export default class HotScreen extends React.Component {
+    render() {
+        return (
+            <View>
+                <Text>This is the Hot Screen</Text>
+            </View>
+        )
+    }
+}
