+++ yekyok/screens/home_screens/NewScreen.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {View} from 'react-native';
+
+export default class NewScreen extends React.Component {
+    render() {
+        return (
+            <View>
+                <Text>This is the New Screen</Text>
+            </View>
+        )
+    }
+}
